requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.3
PyMySQL==1.1.1